+++ src/imp/mod.rs
@@ -37,6 +37,18 @@ mod aarch64;
 #[path = "atomic128/x86_64.rs"]
 mod x86_64;
 
+// Miri and Sanitizer do not support inline assembly.
+#[cfg(all(any(miri, portable_atomic_sanitize_thread), portable_atomic_llvm15))]
+#[cfg(portable_atomic_asm_experimental_arch)]
+#[cfg(any(
+    target_endian = "little",
+    target_feature = "quadword-atomics",
+    portable_atomic_target_feature = "quadword-atomics"
+))]
+#[cfg(target_arch = "powerpc64")]
+#[path = "atomic128/intrinsics.rs"]
+mod powerpc64;
+#[cfg(not(all(any(miri, portable_atomic_sanitize_thread), portable_atomic_llvm15)))]
 #[cfg(portable_atomic_asm_experimental_arch)]
 #[cfg(any(
     target_endian = "little",